requirements.txt
PySide6>=6.6.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
PyMuPDF>=1.23.0
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
mecab-python3>=1.0.5
unidic-lite>=1.0.8
jupyter>=1.0.0
ipykernel>=6.25.0
